Backend/requirements.txt
Flask
Flask-Cors
tensorflow
numpy
scikit-learn
Pillow
